create setup.py for use with distutils, pip, etc
MANIFEST is moved out of the way because it interferes with the setup
function's data_files parameter. It is used from its new location
(MANIFEST.list) to generate the value for the data_files parameter.

An 'asciidoc' shell script is introduced because distutils copies the
files in its scripts parameter and asciidoc.py needs to know its
installed path to access conf files and filters. The script uses the
asciidoc module in the PYTHONPATH which solves this problem.